+++ Sources/SwifterSwift/SwiftStdlib/DoubleExtensions.swift
@@ -52,15 +52,3 @@ func ** (lhs: Double, rhs: Double) -> Double {
     // http://nshipster.com/swift-operators/
     return pow(lhs, rhs)
 }
-
-// swiftlint:disable identifier_name
-prefix operator √
-/// SS: 二次根号
-///
-/// - Parameter double: 开根号的数
-/// - Returns: 平方根
-public prefix func √ (double: Double) -> Double {
-    // http://nshipster.com/swift-operators/
-    return sqrt(double)
-}
-// swiftlint:enable identifier_name