+++ Sources/SwifterSwift/UIKit/UIViewExtensions.swift
@@ -568,7 +568,57 @@ public extension SwifterSwift where Base: UIView {
             base.frame.size.height = newValue
         }
     }
+}
 
+// MARK: - Constraints
+public extension UIView {
+    /// SS: Search constraints until we find one for the given view
+    /// and attribute. This will enumerate ancestors since constraints are
+    /// always added to the common ancestor.
+    ///
+    /// - Parameter attribute: the attribute to find
+    /// - Parameter at: the view to find
+    /// - Returns: matching constraint
+    func findConstraint(attribute: NSLayoutConstraint.Attribute, for view: UIView) -> NSLayoutConstraint? {
+        let constraint = constraints.first {
+            ($0.firstAttribute == attribute && $0.firstItem as? UIView == view) ||
+            ($0.secondAttribute == attribute && $0.secondItem as? UIView == view)
+        }
+        return constraint ?? superview?.findConstraint(attribute: attribute, for: view)
+    }
+
+    /// SS: First width constraint for this view
+    var widthConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .width, for: self)
+    }
+
+    /// SS: First height constraint for this view
+    var heightConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .height, for: self)
+    }
+
+    /// SS: First leading constraint for this view
+    var leadingConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .leading, for: self)
+    }
+
+    /// SS: First trailing constraint for this view
+    var trailingConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .trailing, for: self)
+    }
+
+    /// SS: First top constraint for this view
+    var topConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .top, for: self)
+    }
+
+    /// SS: First bottom constraint for this view
+    var bottomConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .bottom, for: self)
+    }
+}
+
+extension SwifterSwift where Base: UIView {
     /// SS: view 的 x 坐标
     var x: CGFloat {
         get {